Update README.md
Hi Ed,
I'm a WW QuickSight specialist SA. We got notified by our Japan Sales specialist about an issue with this sample.
"This is a template to federate Cognito users to QS but we had feedback from several customers like Hitachi and FP Partner that a single QS user can be shared with all users on Cognito if they use this template.
It seems like the same RoleSessionName is applied using getCredentialsForIdentity, but an appropriate implication is to apply unique RoleSessionName using assumeRoleWithWebIdentity."

I went through the js file and didn't see RoleSessionName being passed. Cognito documentation has this as an optional parameter with default value of 'web-identity'.
This is probably the value getting used here and seems to be in line with the customer feedback.
I have added a note to the Readme to alert subsequent readers of the need to specify RoleSessionName if they decide to use this code as their starting point.
Kindly accept the proposed change. Thanks.

Regards,
Arun Santhosh